Now builds well enough to run in browser
Returning a pointer to a string is much easier to handle in javascript
than passing a pointer as an argument and using that, so switch to that
for analysis_as_json().

Also set MORPHLIB appropriately, as it won't be set in the js
environment.

Also make a smaller stemlib which doesn't include the src to use with
the js version.

Finally, use emscripten's Module.cwrap() to expose the analysis_as_json()
function directly to the javascript.